pyproject.toml
[project]
name = "patent-translation-system"
version = "1.0.0"
description = "특허 번역 자동화 시스템 - Gemini API 기반"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-generativeai>=0.5.0",
    "python-docx>=1.1.0",
    "PyQt6>=6.6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.1",
]

[project.urls]
Homepage = "https://github.com/your-username/patent-translation-system"
Issues = "https://github.com/your-username/patent-translation-system/issues"

[project.optional-dependencies]
dev = [
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
